+++ common-custom-user-data-gradle-plugin/src/main/java/com/gradle/Utils.java
@@ -15,6 +15,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.io.UnsupportedEncodingException;
+import java.net.URI;
 import java.net.URLEncoder;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
@@ -35,8 +36,10 @@ final class Utils {
 
     static Optional<String> projectProperty(String name, ProviderFactory providers, Gradle gradle) {
         if (isGradle65OrNewer()) {
-            Provider<String> property = providers.provider(() -> (String) gradle.getRootProject().findProperty(name)).forUseAtConfigurationTime();
-            return Optional.ofNullable(property.getOrNull());
+            // invalidate configuration cache if different Gradle property value is set on the cmd line,
+            // but in any case access Gradle property directly since project properties set in a build script or
+            // init script are not fetched by ProviderFactory.gradleProperty
+            providers.gradleProperty(name).forUseAtConfigurationTime();
         }
         return Optional.ofNullable((String) gradle.getRootProject().findProperty(name));
     }
@@ -77,12 +80,21 @@ final class Utils {
         return value != null && !value.isEmpty();
     }
 
+    static String stripPrefix(String prefix, String string) {
+        return string.startsWith(prefix) ? string.substring(prefix.length()) : string;
+    }
+
     static String appendIfMissing(String str, String suffix) {
         return str.endsWith(suffix) ? str : str + suffix;
     }
 
-    static String stripPrefix(String prefix, String string) {
-        return string.startsWith(prefix) ? string.substring(prefix.length()) : string;
+    static URI appendPathAndTrailingSlash(URI baseUri, String path) {
+        if (isNotEmpty(path)) {
+            String normalizedBasePath = appendIfMissing(baseUri.getPath(), "/");
+            String normalizedPath = appendIfMissing(stripPrefix("/", path), "/");
+            return baseUri.resolve(normalizedBasePath).resolve(normalizedPath);
+        }
+        return baseUri;
     }
 
     static String urlEncode(String str) {